test: moved tests for own directories
