+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "mocha test/**/*.test.js",
+    "test": "mocha --recursive \"src/**/*.test.js\"",
     "watch:test": "npm run test -- --watch",
     "lint": "./node_modules/.bin/eslint src/**/*.js"
   },
